refactor(edge): split h3 edge integration tests
