+++ crates/edge/src/quic_listener/tests.rs
@@ -1,4 +1,3 @@
-
 use std::{
     collections::HashMap,
     path::Path,
