+++ crates/edge/tests/h3_edge.rs
@@ -17,6 +17,8 @@ use rcgen::{Certificate, CertificateParams, SanType};
 use tempfile::{TempDir, tempdir};
 use tokio::net::TcpListener;
 
+mod support;
+
 use spooky_config::config::{
     Backend, ClientAuth, Config, HealthCheck, Listen, LoadBalancing, Log, LogFormat, Security, Tls,
     UpstreamTls,
@@ -27,26 +29,7 @@ use spooky_edge::constants::{
     QUIC_INITIAL_MAX_STREAMS_BIDI, QUIC_INITIAL_MAX_STREAMS_UNI, QUIC_INITIAL_STREAM_DATA,
     REQUEST_TIMEOUT_SECS, UDP_READ_TIMEOUT_MS,
 };
-
-fn local_listener_bind_available() -> bool {
-    let tcp_ok = match std::net::TcpListener::bind("127.0.0.1:0") {
-        Ok(listener) => {
-            drop(listener);
-            true
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::PermissionDenied => false,
-        Err(_) => true,
-    };
-    let udp_ok = match std::net::UdpSocket::bind("127.0.0.1:0") {
-        Ok(socket) => {
-            drop(socket);
-            true
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::PermissionDenied => false,
-        Err(_) => true,
-    };
-    tcp_ok && udp_ok
-}
+use support::net::local_listener_bind_available;
 
 fn write_test_certs(dir: &TempDir) -> (String, String) {
     let mut params = CertificateParams::new(vec!["localhost".into()]);
@@ -514,108 +497,11 @@ fn assert_cid_sync_invariants(listener: &QUICListener) {
     }
 }
 
-#[test]
-fn http3_request_is_accepted_and_parsed() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-    let config = make_config(0, cert, key, "127.0.0.1:1".to_string());
-    let mut listener = QUICListener::new(config).expect("failed to create listener");
-    let addr = listener.socket.local_addr().unwrap();
+#[path = "h3_edge/startup.rs"]
+mod startup;
 
-    let stop = Arc::new(AtomicBool::new(false));
-    let stop_flag = stop.clone();
-
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let handle = rt.spawn_blocking(move || {
-        while !stop_flag.load(Ordering::Relaxed) {
-            listener.poll();
-        }
-    });
-
-    let body = run_h3_client(addr).expect("client request failed");
-    stop.store(true, Ordering::Relaxed);
-    handle.abort();
-
-    assert!(body.contains("upstream error"));
-}
-
-#[test]
-fn invalid_backend_scheme_is_rejected_at_startup() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-    let config = make_config(0, cert, key, "ftp://127.0.0.1:8080".to_string());
-    match QUICListener::new(config) {
-        Ok(_) => panic!("invalid backend scheme should fail startup"),
-        Err(err) => {
-            assert!(
-                err.to_string().contains("backend_address_invalid"),
-                "unexpected startup error: {err}"
-            );
-        }
-    }
-}
-
-#[test]
-fn server_rotates_scids_for_active_connection() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("failed to create temp dir");
-    let (cert, key) = write_test_certs(&dir);
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend("ok\n"));
-    let config = make_config(0, cert, key, backend_addr.to_string());
-    let listener = Arc::new(Mutex::new(
-        QUICListener::new(config).expect("failed to create listener"),
-    ));
-    let addr = listener
-        .lock()
-        .expect("listener lock")
-        .socket
-        .local_addr()
-        .unwrap();
-
-    let stop = Arc::new(AtomicBool::new(false));
-    let stop_flag = stop.clone();
-    let listener_task = listener.clone();
-
-    let handle = thread::spawn(move || {
-        while !stop_flag.load(Ordering::Relaxed) {
-            if let Ok(mut guard) = listener_task.lock() {
-                guard.poll();
-            }
-        }
-    });
-
-    let (max_spare_dcids, completed_requests) =
-        run_h3_client_multiple_requests(addr, 12).expect("client requests failed");
-
-    stop.store(true, Ordering::Relaxed);
-    let _ = handle.join();
-
-    let listener_guard = listener.lock().expect("listener lock");
-    let rotations = listener_guard
-        .metrics
-        .scid_rotations
-        .load(Ordering::Relaxed);
-
-    assert!(
-        max_spare_dcids > 0,
-        "client never observed additional destination CIDs"
-    );
-    assert!(
-        completed_requests > 0,
-        "client did not complete any request"
-    );
-    assert!(rotations > 0, "server did not rotate any SCID");
-    assert_cid_sync_invariants(&listener_guard);
-}
+#[path = "h3_edge/scid.rs"]
+mod scid;
 
 // ---------------------------------------------------------------------------
 // Malformed packet hardening tests (task 1.1)
@@ -661,179 +547,8 @@ fn assert_maps_empty(listener: &QUICListener) {
     );
 }
 
-/// A single zero-byte UDP datagram must be dropped without panic and leave all
-/// maps empty.
-#[test]
-fn malformed_zero_length_datagram_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    send_udp(addr, &[]);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// A single-byte payload cannot be a valid QUIC header; listener must drop it.
-#[test]
-fn malformed_single_byte_datagram_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    send_udp(addr, &[0xFF]);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// Completely random garbage bytes must not panic and must leave maps clean.
-#[test]
-fn malformed_random_garbage_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    let garbage: Vec<u8> = (0u8..64).collect();
-    send_udp(addr, &garbage);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// A valid-looking QUIC long header with a plausible type byte but truncated
-/// body should be rejected cleanly.
-#[test]
-fn malformed_truncated_long_header_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    // Long-header first byte: version-specific Initial packet marker (0xC0 | 0x00)
-    // followed by the QUIC version, then truncated before DCIL/SCIL fields.
-    let truncated: &[u8] = &[
-        0xC0, // long-header flag + Initial type bits
-        0x00, 0x00, 0x00,
-        0x01, // QUIC v1
-              // deliberately truncated here (no DCIL/SCIL/lengths)
-    ];
-    send_udp(addr, truncated);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// A long header with a DCID length that overflows the packet should be
-/// rejected without panicking.
-#[test]
-fn malformed_dcid_length_overflow_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    // Craft a packet whose DCID length field claims 255 bytes but the packet
-    // ends immediately after.
-    let mut pkt = vec![
-        0xC0, // long-header Initial
-        0x00, 0x00, 0x00, 0x01, // QUIC v1
-        0xFF, // DCID length = 255 (but no bytes follow)
-    ];
-    // pad with some bytes but far fewer than 255
-    pkt.extend_from_slice(&[0xAB; 8]);
-    send_udp(addr, &pkt);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// A short-header packet destined for an unknown connection must be silently
-/// dropped; it must not create a new connection entry.
-#[test]
-fn short_header_unknown_connection_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    // Short header: first bit 0, remaining bits arbitrary. Use a 20-byte DCID
-    // that does not correspond to any established connection.
-    let mut pkt = vec![0x40u8]; // short-header flag (bit 7 = 0, bit 6 = 1 for fixed bit)
-    pkt.extend_from_slice(&[
-        0xDE, 0xAD, 0xBE, 0xEF, 0xCA, 0xFE, 0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08,
-        0x09, 0x0A, 0x0B, 0x0C, 0x0D,
-    ]);
-    send_udp(addr, &pkt);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// A Retry packet from an unknown peer must be ignored without creating state.
-#[test]
-fn retry_packet_for_unknown_connection_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    // Long-header Retry type: bits 0xF0 with version and minimal fields.
-    // quiche will parse the header but the listener should not create a conn.
-    let mut pkt = vec![
-        0xF0, // long-header, Retry type bits
-        0x00, 0x00, 0x00, 0x01, // QUIC v1
-        0x08, // DCID len = 8
-    ];
-    pkt.extend_from_slice(&[0x11; 8]); // DCID
-    pkt.push(0x08); // SCID len = 8
-    pkt.extend_from_slice(&[0x22; 8]); // SCID
-    // Retry token (arbitrary, no integrity tag)
-    pkt.extend_from_slice(&[0x99; 16]);
-    send_udp(addr, &pkt);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// A Handshake packet for which no connection exists must be dropped cleanly.
-#[test]
-fn handshake_packet_unknown_connection_is_dropped() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-    // Long-header Handshake type: 0xE0
-    let mut pkt = vec![
-        0xE0, // long-header, Handshake type bits
-        0x00, 0x00, 0x00, 0x01, // QUIC v1
-        0x08, // DCID len = 8
-    ];
-    pkt.extend_from_slice(&[0x33; 8]); // DCID
-    pkt.push(0x08); // SCID len = 8
-    pkt.extend_from_slice(&[0x44; 8]); // SCID
-    // Packet number + payload (garbage)
-    pkt.extend_from_slice(&[0x00, 0x00, 0x00, 0x00]);
-    pkt.extend_from_slice(&[0xAA; 20]);
-    send_udp(addr, &pkt);
-    listener.poll();
-    assert_maps_empty(&listener);
-}
-
-/// Repeated bursts of malformed packets must not accumulate any routing state.
-#[test]
-fn repeated_malformed_packets_leave_maps_consistent() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let (mut listener, addr) = make_listener();
-
-    let payloads: &[&[u8]] = &[
-        &[],                                   // zero-length
-        &[0xFF],                               // single byte
-        &[0x00; 16],                           // all-zero short
-        &[0xFF; 64],                           // all-ones garbage
-        &[0xC0, 0x00, 0x00, 0x00, 0x01, 0xFF], // truncated long header
-        &[0x40, 0xDE, 0xAD, 0xBE, 0xEF, 0x00], // short header, unknown DCID
-    ];
-
-    for payload in payloads {
-        send_udp(addr, payload);
-        listener.poll();
-    }
-
-    assert_maps_empty(&listener);
-}
+#[path = "h3_edge/malformed.rs"]
+mod malformed;
 
 // ---------------------------------------------------------------------------
 // Connection flood / rate-limit tests (task 1.2)
@@ -950,183 +665,8 @@ fn build_initial_packet(dest_addr: std::net::SocketAddr) -> Vec<u8> {
     out
 }
 
-/// When burst=1 and rate is near-zero, the first Initial packet creates a
-/// connection, and subsequent ones in the same instant are dropped.
-#[test]
-fn connection_flood_is_rate_limited() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    // burst=1, rate=1/s: after the first accept the bucket is empty.
-    let config = make_config_with_rate_limit(0, cert, key, "127.0.0.1:1".to_string(), 1, 1);
-    let mut listener = QUICListener::new(config).expect("listener");
-    let addr = listener.socket.local_addr().unwrap();
-
-    // Send enough distinct Initial packets to saturate the rate limit.
-    // Each packet comes from a different SCID so the listener sees each as a
-    // new connection attempt (no existing DCID match).
-    const FLOOD_COUNT: usize = 10;
-    // Build all packets up front so token consumption happens in a tight burst.
-    // This avoids slow-host timing where packet construction can allow a 1/s
-    // refill and make the assertion flaky.
-    let packets: Vec<Vec<u8>> = (0..FLOOD_COUNT)
-        .map(|_| build_initial_packet(addr))
-        .collect();
-    for pkt in &packets {
-        send_udp(addr, pkt);
-    }
-    for _ in 0..FLOOD_COUNT {
-        listener.poll();
-    }
-
-    // With burst=1 only the very first packet can create a connection.
-    // All subsequent ones are dropped by the rate limiter.
-    assert!(
-        listener.connections().len() <= 1,
-        "rate limiter must cap connections at burst=1, got {}",
-        listener.connections().len()
-    );
-}
-
-/// Hard cap on active connections should reject additional Initial packets
-/// even when token-bucket rate limits are permissive.
-#[test]
-fn active_connection_cap_rejects_excess_initial_packets() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    let mut config =
-        make_config_with_rate_limit(0, cert, key, "127.0.0.1:1".to_string(), 10_000, 10_000);
-    config.performance.max_active_connections = 1;
-    let mut listener = QUICListener::new(config).expect("listener");
-    let addr = listener.socket.local_addr().unwrap();
-
-    const FLOOD_COUNT: usize = 8;
-    for _ in 0..FLOOD_COUNT {
-        let pkt = build_initial_packet(addr);
-        send_udp(addr, &pkt);
-        listener.poll();
-    }
-
-    assert!(
-        listener.connections().len() <= 1,
-        "active connection cap must keep at most one connection, got {}",
-        listener.connections().len()
-    );
-    assert!(
-        listener
-            .metrics
-            .connection_cap_rejects
-            .load(Ordering::Relaxed)
-            > 0,
-        "connection cap should emit rejection metrics"
-    );
-}
-
-/// Once draining starts, unknown/new Initial packets must not create a
-/// connection, even if admission limits would otherwise allow it.
-#[test]
-fn draining_mode_rejects_initial_when_no_connections_exist() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    let config = make_config_with_rate_limit(0, cert, key, "127.0.0.1:1".to_string(), 10_000, 20);
-    let mut listener = QUICListener::new(config).expect("listener");
-    let addr = listener.socket.local_addr().unwrap();
-
-    listener.start_draining();
-
-    let pkt = build_initial_packet(addr);
-    send_udp(addr, &pkt);
-    listener.poll();
-
-    assert_eq!(
-        listener.connections().len(),
-        0,
-        "no new connection should be admitted after drain starts"
-    );
-}
-
-/// Draining should preserve existing connection processing but reject any new
-/// connection admission from unknown Initial packets.
-#[test]
-fn draining_mode_rejects_new_initial_after_existing_connection() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    let config = make_config_with_rate_limit(0, cert, key, "127.0.0.1:1".to_string(), 10_000, 20);
-    let mut listener = QUICListener::new(config).expect("listener");
-    let addr = listener.socket.local_addr().unwrap();
-
-    let first = build_initial_packet(addr);
-    send_udp(addr, &first);
-    listener.poll();
-    assert_eq!(
-        listener.connections().len(),
-        1,
-        "first Initial should create a baseline connection before drain"
-    );
-    let known_connection_ids: std::collections::HashSet<Vec<u8>> = listener
-        .connections()
-        .keys()
-        .map(|cid| cid.to_vec())
-        .collect();
-
-    listener.start_draining();
-
-    for _ in 0..5 {
-        let pkt = build_initial_packet(addr);
-        send_udp(addr, &pkt);
-        listener.poll();
-    }
-
-    let post_drain_ids: std::collections::HashSet<Vec<u8>> = listener
-        .connections()
-        .keys()
-        .map(|cid| cid.to_vec())
-        .collect();
-    assert!(
-        post_drain_ids.is_subset(&known_connection_ids),
-        "draining mode must reject unknown/new Initial packets (before={known_connection_ids:?}, after={post_drain_ids:?})"
-    );
-}
-
-/// Normal traffic well below the rate limit must not be affected.
-/// With a generous burst and rate, all N connection attempts succeed.
-#[test]
-fn normal_traffic_below_rate_limit_is_unaffected() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    // burst=20, rate=10000/s: far above what we'll send.
-    let config = make_config_with_rate_limit(0, cert, key, "127.0.0.1:1".to_string(), 10_000, 20);
-    let mut listener = QUICListener::new(config).expect("listener");
-    let addr = listener.socket.local_addr().unwrap();
-
-    const REQUEST_COUNT: usize = 5;
-    for _ in 0..REQUEST_COUNT {
-        let pkt = build_initial_packet(addr);
-        send_udp(addr, &pkt);
-        listener.poll();
-    }
-
-    assert_eq!(
-        listener.connections().len(),
-        REQUEST_COUNT,
-        "all {} connections should be accepted when below rate limit",
-        REQUEST_COUNT
-    );
-}
+#[path = "h3_edge/admission.rs"]
+mod admission;
 
 // ---------------------------------------------------------------------------
 // Connection-lifecycle churn stress test (task 4.3)
@@ -1247,89 +787,5 @@ fn stress_close_gracefully(socket: &UdpSocket, conn: &mut quiche::Connection) {
     }
 }
 
-/// Stress test: repeatedly connect, optionally complete a request, then close
-/// or abruptly drop.  After all rounds the listener maps must be fully empty.
-#[test]
-fn lifecycle_churn_leaves_no_orphaned_state() {
-    if !local_listener_bind_available() {
-        return;
-    }
-    const ROUNDS: usize = 30;
-    // Idle timeout matches the client config (500 ms).
-    const IDLE_MS: u64 = 500;
-
-    let dir = tempdir().expect("tempdir");
-    let (cert, key) = write_test_certs(&dir);
-    let rt = tokio::runtime::Runtime::new().expect("runtime");
-    let backend_addr = rt.block_on(start_h2_backend("ok\n"));
-
-    let mut config = make_config(0, cert, key, backend_addr.to_string());
-    // Match the client-side idle timeout so both sides agree.
-    config.performance.quic_max_idle_timeout_ms = IDLE_MS;
-    // Generous limits so the stress loop is never rate-limited.
-    config.performance.new_connections_per_sec = 10_000;
-    config.performance.new_connections_burst = 1_000;
-
-    let listener = Arc::new(Mutex::new(
-        QUICListener::new(config).expect("failed to create listener"),
-    ));
-    let server_addr = listener.lock().unwrap().socket.local_addr().unwrap();
-
-    let stop = Arc::new(AtomicBool::new(false));
-    let stop_flag = stop.clone();
-    let listener_poll = listener.clone();
-    let poll_handle = thread::spawn(move || {
-        while !stop_flag.load(Ordering::Relaxed) {
-            if let Ok(mut g) = listener_poll.lock() {
-                g.poll();
-            }
-        }
-    });
-
-    for round in 0..ROUNDS {
-        let (socket, _local_addr, mut conn, _h3) =
-            stress_connect(server_addr).unwrap_or_else(|e| panic!("round {round}: connect: {e}"));
-
-        socket
-            .set_read_timeout(Some(Duration::from_millis(100)))
-            .expect("set_read_timeout");
-
-        // Even rounds close gracefully; odd rounds abruptly drop the socket.
-        if round % 2 == 0 {
-            stress_close_gracefully(&socket, &mut conn);
-        }
-        // Odd rounds: drop socket and conn; server detects via idle timeout.
-    }
-
-    // Allow the idle timeout (500 ms) to fire and the listener to sweep all
-    // timed-out connections.  We wait 3× idle timeout for safety margin.
-    thread::sleep(Duration::from_millis(IDLE_MS * 3));
-
-    stop.store(true, Ordering::Relaxed);
-    let _ = poll_handle.join();
-
-    // Drive one final poll to flush any remaining close/sweep work.
-    if let Ok(mut g) = listener.lock() {
-        for _ in 0..20 {
-            g.poll();
-        }
-    }
-
-    let guard = listener.lock().expect("listener lock");
-    assert_cid_sync_invariants(&guard);
-    assert!(
-        guard.connections().is_empty(),
-        "connections must be empty after churn, found {} entries",
-        guard.connections().len()
-    );
-    assert!(
-        guard.cid_routes().is_empty(),
-        "cid_routes must be empty after churn, found {} entries",
-        guard.cid_routes().len()
-    );
-    assert!(
-        guard.peer_routes().is_empty(),
-        "peer_routes must be empty after churn, found {} entries",
-        guard.peer_routes().len()
-    );
-}
+#[path = "h3_edge/lifecycle.rs"]
+mod lifecycle;
